Update Plugin API after sync between SDKs

+++ Runtime/Api/BidonReward.cs
@@ -1,7 +1,8 @@
+// ReSharper disable CheckNamespace
+
 using System;
 using System.Diagnostics.CodeAnalysis;
 
-// ReSharper disable once CheckNamespace
 namespace Bidon.Mediation
 {
     [Serializable]
@@ -12,7 +13,7 @@ namespace Bidon.Mediation
     {
         public string Label;
 
-        public int Amount;
+        public double Amount;
 
         public string ToJsonString(bool isPretty = false) => UnityEngine.JsonUtility.ToJson(this, isPretty);
     }
